mmu project
MMU is a hardware component responsible for memory management operations. Its purpose is to move pages from the main memory to the secondary memory .
The operating system stores the data in memory through pages and manages the memory with an action called paging.
This uses an algorithm which, with a certain logic, selects which page on the system to transfer from the main memory to the secondary when the main memory is full.

In this project, we introduced the mmu operation using the principles of oop, encapsulation, design patterns.

The rar file has a link to youtube which I illustrate the code.